backend/requirements.txt
fastapi>=0.110,<1.0
firebase-admin>=6.5,<8.0
google-auth>=2.30,<3.0
httpx>=0.27,<1.0
openai>=2.0,<3.0
python-dotenv>=1.0,<2.0
uvicorn[standard]>=0.27,<1.0
